Fix imports and renames of the databases constructors
